Refactor participant ID retrieval to use session-based authentication across multiple servlets

+++ src/main/java/com/example/jee_event_manager/model/Inscription.java
@@ -37,7 +37,7 @@ public class Inscription {
     private Participant participant;
     
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "evenement_id", nullable = false, columnDefinition = "INTEGER")
+    @JoinColumn(name = "evenement_id", nullable = false, columnDefinition = "BIGINT")
     private Evenement evenement;
     
     @Column(name = "type_billet", length = 50)
